--- src/main/java/fleetManager/FleetFileParser.java
@@ -1,4 +1,0 @@
-package fleetManager;
-
-public class FleetFileParser {
-}
