+++ src/main/java/fleetManager/FleetManagerImplementation.java
@@ -1,0 +1,34 @@
+package fleetManager;
+
+import fleetManager.hosts.Host;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class FleetManagerImplementation implements FleetManager {
+//singleton
+    public FleetManagerImplementation(){
+
+    }
+
+    public void addHost(Host host){
+
+    }
+
+    public void removeHost(Host host){
+
+    }
+
+    public void enableHost(Host host){
+
+    }
+
+    public void disableHost(Host host){
+
+    }
+
+    public List<Host> getHosts(){
+        List<Host> hosts = new ArrayList<Host>();
+        return hosts;
+    }
+}
