+++ src/main/java/fleetManager/HealthCheck.java
@@ -1,0 +1,7 @@
+package fleetManager;
+
+import fleetManager.hosts.Host;
+
+public interface HealthCheck {
+
+}
