+++ src/main/java/fleetManager/HeathCheckImplementation.java
@@ -1,0 +1,14 @@
+package fleetManager;
+
+import fleetManager.command.AddHostCommand;
+import fleetManager.command.DisableHostCommand;
+import fleetManager.command.EnableHostCommand;
+import fleetManager.command.RemoveHostCommand;
+import fleetManager.hosts.Host;
+
+public class HeathCheckImplementation implements HealthCheck {
+
+    HeathCheckImplementation(){
+
+    }
+}
