--- src/main/java/fleetManager/IFleetManager.java
@@ -1,11 +1,0 @@
-package fleetManager;
-
-import fleetManager.hosts.Host;
-
-public interface IFleetManager {
-    void addHost(Host host);
-    void removeHost(Host host);
-    void enableHost(Host host);
-    void disableHost(Host host);
-    void getHosts();
-}
