+++ src/main/java/fleetManager/command/EnableHostCommand.java
@@ -1,19 +1,19 @@
 package fleetManager.command;
 
 import fleetManager.hosts.Host;
-import fleetManager.hosts.HostManager;
+import fleetManager.FleetManagerImplementation;
 
 public class EnableHostCommand implements Command {
     private Host host;
-    private HostManager hostManager;
+    private FleetManagerImplementation fleetManagerImplementation;
 
     public EnableHostCommand(Host host){
-        hostManager=new HostManager();
+        fleetManagerImplementation =new FleetManagerImplementation();
         this.host=host;
     }
 
     public void execute() {
-        hostManager.enableHost(host);
+        fleetManagerImplementation.enableHost(host);
     }
 
 }
