+++ src/main/java/fleetManager/command/GetHostsCommand.java
@@ -1,18 +1,19 @@
 package fleetManager.command;
 
-import fleetManager.hosts.Host;
 import fleetManager.FleetManagerImplementation;
+import fleetManager.hosts.Host;
 
-public class RemoveHostCommand implements Command {
-    private Host host;
+import java.util.List;
+
+public class GetHostsCommand implements Command {
     private FleetManagerImplementation fleetManagerImplementation;
+    private List<Host> hosts;
 
-    public RemoveHostCommand(Host host){
+    public GetHostsCommand(){
         fleetManagerImplementation =new FleetManagerImplementation();
-        this.host=host;
     }
 
     public void execute() {
-        fleetManagerImplementation.removeHost(host);
+        hosts=fleetManagerImplementation.getHosts();
     }
 }
