--- src/main/java/fleetManager/hosts/HostManager.java
@@ -1,26 +1,0 @@
-package fleetManager.hosts;
-
-public class HostManager implements IHostManger{
-
-    public HostManager(){
-
-    }
-
-    public void addHost(Host host){
-
-    }
-
-    public void removeHost(Host host){
-
-    }
-
-    public void enableHost(Host host){
-
-    }
-
-    public void disableHost(Host host){
-
-    }
-
-
-}
