--- src/main/java/fleetManager/hosts/IHostManger.java
@@ -1,8 +1,0 @@
-package fleetManager.hosts;
-
-public interface IHostManger {
-    void addHost(Host host);
-    void removeHost(Host host);
-    void enableHost(Host host);
-    void disableHost(Host host);
-}
